중간 진행중

+++ mes_back/src/main/java/com/mes_back/repository/OrderOutboundRepository.java
@@ -3,5 +3,6 @@ package com.mes_back.repository;
 import com.mes_back.entity.OrderOutbound;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderOutboundRepository extends JpaRepository<OrderOutbound,Integer> {
+// ✅ (수정) JpaRepository<OrderOutbound, Integer> -> JpaRepository<OrderOutbound, Long>
+public interface OrderOutboundRepository extends JpaRepository<OrderOutbound, Long> {
 }
